Work is still in progress

+++ lib/Microphone/Microphone.cpp
@@ -10,9 +10,10 @@ void Microphone::initPin() const {
     pinMode(this->pin, INPUT);
 }
 
-bool Microphone::isTriggered(unsigned short treshold, unsigned short streak) {
+bool Microphone::isTriggered(unsigned short treshold, uint8_t streak) {
     int reading = analogRead(this->pin);
-    if (reading >= static_cast<int>(treshold)) {
+    if (reading >=
+        treshold) { // note: comparing signed value (reading) and unsigned (treshold) because we are 100% sure, that reading will never be negative
         this->trigger_streak += 1;
         if (this->trigger_streak == streak) {
             this->trigger_streak = 0;
